+++ biblioteca-4/comunes/biblioteca-mysql.php
@@ -68,7 +68,6 @@ function conectaDb()
         $tmp->exec("set names utf8mb4");
         return $tmp;
     } catch (PDOException $e) {
-        cabecera("Error grave", MENU_VOLVER, PROFUNDIDAD_1);
         print "    <p class=\"aviso-error\">Error: No puede conectarse con la base de datos.</p>\n";
         print "\n";
         print "    <p class=\"aviso-error\">Error: " . $e->getMessage() . "</p>\n";